% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{get_price_data}
\alias{get_price_data}
\title{Scraping price data from SNIIM
The function webscrape the tables containing fish price data from the SNIIM using
year, month, starting day (day1), and last day (day2) that is wanted. The structure of the website for now
only allows single year and month queries, therefore it needs to be looped in some whay to be obtained.
Sometimes the URL returns an error depending on the dates interted. This is caused by the somewhat inestable structure
of the website. Using different dates might solve the issue. By default it scrapes today's prices.}
\usage{
get_price_data(
  year = as.numeric(lubridate::year(Sys.Date())),
  month = as.numeric(lubridate::month(Sys.Date())),
  day1 = as.numeric(lubridate::day(Sys.Date() - 1)),
  day2 = as.numeric(lubridate::day(Sys.Date()))
)
}
\arguments{
\item{year}{the year input to scrape, must be numeric, defaults to current year}

\item{month}{the month input to scrape, must be numeric, defaults to current month}

\item{day1}{the starting day input to scrape, must be numeric, defaults to current day}

\item{day2}{the final day input to scrape, must be numeric, defaults to current day, can be different than day1}
}
\value{
a data.frame object
}
\description{
Scraping price data from SNIIM
The function webscrape the tables containing fish price data from the SNIIM using
year, month, starting day (day1), and last day (day2) that is wanted. The structure of the website for now
only allows single year and month queries, therefore it needs to be looped in some whay to be obtained.
Sometimes the URL returns an error depending on the dates interted. This is caused by the somewhat inestable structure
of the website. Using different dates might solve the issue. By default it scrapes today's prices.
}
\examples{
# Scraping prices from a custom date
\dontrun{
get_price_data(year = 2022, month = 02, day1 = 01, day2 = 22)
}
}
