% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_correction.R
\docType{data}
\name{inflation_correction}
\alias{inflation_correction}
\title{Inflation correction}
\format{
a data.frame with 23 observations and 2 variables
\describe{
\item{year}{year of reference for the inflation value}
\item{amount}{correction value for the inflation}
...
}
}
\source{
CBMC
}
\usage{
inflation_correction
}
\description{
A data.frame table that has the corretion index for price inflation up to 2000
}
\keyword{datasets}
