% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytnyt.R
\name{nytnyt}
\alias{nytnyt}
\title{Night Night function
The function create random time gaps bewtween scraping requests to site to simulate normal traffic behavior and not saturate the site.}
\usage{
nytnyt(periods = c(1, 1.5))
}
\arguments{
\item{periods}{default waiting periods interval that will be randomly chosen}
}
\value{
message
}
\description{
Night Night function
The function create random time gaps bewtween scraping requests to site to simulate normal traffic behavior and not saturate the site.
}
\examples{
nytnyt()

}
