% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_sniim.R
\name{scraping_sniim}
\alias{scraping_sniim}
\title{Scraping price data website
The function loops into all urls from fishery prices with the dates indicated and gets a list of data.frames}
\usage{
scraping_sniim(weblist)
}
\arguments{
\item{weblist}{a url from the sniim market page to be scraped}
}
\value{
a data.frame
}
\description{
Scraping price data website
The function loops into all urls from fishery prices with the dates indicated and gets a list of data.frames
}
\examples{
\dontrun{
scraping_sniim(url)
}

}
