% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniimscraper-package.R
\docType{package}
\name{sniimscraper-package}
\alias{sniimscraper}
\alias{sniimscraper-package}
\title{sniimscraper: WebScraping Fish Price Data From Mexican Government Sources}
\description{
WebScrape fish data from the Mexican Government's Sistema Nacional de Informacion e Integracion de Mercados (SNIIM). It also features some tailored functions to correct prices for inflation, associate catch data and forecast capabilities.
}
\author{
\strong{Maintainer}: Fabio Favoretto \email{fabio@gocmarineprogram.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
