% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dictionary.R
\docType{data}
\name{sp_dictionary}
\alias{sp_dictionary}
\title{Species names dictionary}
\format{
A data frame of 148 observations and 3 variables
\describe{
\item{product}{the original name from the database}
\item{short_names}{named corrected and simplified}
\item{ID}{a numeric ID per each species name}
...
}
}
\source{
CBMC
}
\usage{
sp_dictionary
}
\description{
A data.frame containing species names correspondences that can be corrected
}
\keyword{datasets}
